+++ workers/ohsome_quality_analyst/api/api.py
@@ -1,9 +1,14 @@
+import json
 import logging
-from typing import Optional
+from typing import Optional, Union
 
 import pydantic
-from fastapi import FastAPI
+from fastapi import FastAPI, Request, status
+from fastapi.encoders import jsonable_encoder
+from fastapi.exceptions import RequestValidationError
 from fastapi.middleware.cors import CORSMiddleware
+from fastapi.responses import JSONResponse
+from pydantic import ValidationError
 
 from ohsome_quality_analyst import (
     __author__,
@@ -33,6 +38,9 @@ from ohsome_quality_analyst.utils.definitions import (
     get_layer_names,
     get_report_names,
 )
+from ohsome_quality_analyst.utils.exceptions import OhsomeApiError, SizeRestrictionError
+
+MEDIA_TYPE_GEOJSON = "application/geo+json"
 
 configure_logging()
 logging.info("Logging enabled")
@@ -59,6 +67,48 @@ app.add_middleware(
 )
 
 
+@app.exception_handler(ValidationError)
+@app.exception_handler(RequestValidationError)
+async def validation_exception_handler(
+    request: Request, exception: Union[RequestValidationError, ValidationError]
+):
+    """Override request validation exceptions.
+
+    `pydantic` raises on exception regardless of the number of errors found.
+    The `ValidationError` will contain information about all the errors.
+
+    FastAPIs `RequestValidationError` is a subclass of pydantics `ValidationError`.
+    Because of the usage of `@pydantic.validate_arguments` decorator
+    `ValidationError` needs to be specified in this handler as well.
+    """
+    return JSONResponse(
+        status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
+        content=jsonable_encoder(
+            {
+                "apiVersion": __version__,
+                "detail": exception.errors(),
+                "type": "RequestValidationError",
+            },
+        ),
+    )
+
+
+@app.exception_handler(OhsomeApiError)
+@app.exception_handler(SizeRestrictionError)
+async def oqt_exception_handler(
+    request: Request, exception: Union[OhsomeApiError, SizeRestrictionError]
+):
+    """Exception handler for custom OQT exceptions."""
+    return JSONResponse(
+        status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
+        content={
+            "apiVersion": __version__,
+            "detail": exception.message,
+            "type": exception.name,
+        },
+    )
+
+
 def empty_api_response() -> dict:
     return {
         "apiVersion": __version__,
@@ -69,7 +119,7 @@ def empty_api_response() -> dict:
     }
 
 
-@app.get("/indicator/{name}")
+@app.get("/indicator")
 async def get_indicator(
     name: IndicatorEnum,
     layerName: LayerEnum,
@@ -88,7 +138,10 @@ async def get_indicator(
     The Feature properties of the input GeoJSON will be preserved
     if they do not collide with the properties set by OQT.
     """
+    if bpolys is not None:
+        bpolys = json.loads(bpolys)
     raw = {
+        "name": name,
         "layerName": layerName,
         "bpolys": bpolys,
         "dataset": dataset,
@@ -96,12 +149,11 @@ async def get_indicator(
         "fidField": fidField,
     }
     parameters = {k: v for k, v in raw.items() if v is not None}
-    return await _fetch_indicator(name.value, parameters)
+    return await _fetch_indicator(parameters)
 
 
-@app.post("/indicator/{name}")
+@app.post("/indicator")
 async def post_indicator(
-    name: IndicatorEnum,
     parameters: IndicatorRequestModel,
 ):
     """Create an Indicator.
@@ -114,23 +166,22 @@ async def post_indicator(
     The Feature properties of the input GeoJSON will be preserved
     if they do not collide with the properties set by OQT.
     """
-    return await _fetch_indicator(name.value, parameters)
+    return await _fetch_indicator(parameters)
 
 
 @pydantic.validate_arguments
 async def _fetch_indicator(
-    name: str,
     parameters: IndicatorRequestModel,
 ) -> dict:
     p = parameters.dict()
     dataset = p["dataset"]
     fid_field = p["fid_field"]
     if dataset is not None:
-        dataset = p["dataset"].value
+        dataset = dataset.value
     if fid_field is not None:
         fid_field = fid_field.value
     geojson_object = await oqt.create_indicator_as_geojson(
-        name,
+        p["name"].value,
         p["layer_name"].value,
         p["bpolys"],
         dataset,
@@ -140,10 +191,12 @@ async def _fetch_indicator(
     )
     response = empty_api_response()
     response.update(geojson_object)
-    return response
+    return JSONResponse(
+        content=jsonable_encoder(response), media_type=MEDIA_TYPE_GEOJSON
+    )
 
 
-@app.get("/report/{name}")
+@app.get("/report")
 async def get_report(
     name: ReportEnum,
     bpolys: Optional[str] = None,
@@ -161,18 +214,21 @@ async def get_report(
     The Feature properties of the input GeoJSON will be preserved
     if they do not collide with the properties set by OQT.
     """
+    if bpolys is not None:
+        bpolys = json.loads(bpolys)
     raw = {
+        "name": name,
         "bpolys": bpolys,
         "dataset": dataset,
         "featureId": featureId,
         "fidField": fidField,
     }
     parameters = {k: v for k, v in raw.items() if v is not None}
-    return await _fetch_report(name.value, parameters)
+    return await _fetch_report(parameters)
 
 
-@app.post("/report/{name}")
-async def post_report(name: ReportEnum, parameters: ReportRequestModel):
+@app.post("/report")
+async def post_report(parameters: ReportRequestModel):
     """Create a Report.
 
     Either the parameters `dataset` and `feature id` has to be provided
@@ -183,20 +239,20 @@ async def post_report(name: ReportEnum, parameters: ReportRequestModel):
     The Feature properties of the input GeoJSON will be preserved
     if they do not collide with the properties set by OQT.
     """
-    return await _fetch_report(name.value, parameters)
+    return await _fetch_report(parameters)
 
 
 @pydantic.validate_arguments
-async def _fetch_report(name: str, parameters: ReportRequestModel):
+async def _fetch_report(parameters: ReportRequestModel):
     p = parameters.dict()
     dataset = p["dataset"]
     fid_field = p["fid_field"]
     if dataset is not None:
-        dataset = p["dataset"].value
+        dataset = dataset.value
     if fid_field is not None:
         fid_field = fid_field.value
     geojson_object = await oqt.create_report_as_geojson(
-        name,
+        p["name"].value,
         bpolys=p["bpolys"],
         dataset=dataset,
         feature_id=p["feature_id"],
@@ -205,7 +261,9 @@ async def _fetch_report(name: str, parameters: ReportRequestModel):
     )
     response = empty_api_response()
     response.update(geojson_object)
-    return response
+    return JSONResponse(
+        content=jsonable_encoder(response), media_type=MEDIA_TYPE_GEOJSON
+    )
 
 
 @app.get("/regions")
@@ -219,9 +277,12 @@ async def get_available_regions(asGeoJSON: bool = False):
     if asGeoJSON is True:
         regions = await db_client.get_regions_as_geojson()
         response.update(regions)
+        return JSONResponse(
+            content=jsonable_encoder(response), media_type=MEDIA_TYPE_GEOJSON
+        )
     else:
         response["result"] = await db_client.get_regions()
-    return response
+        return response
 
 
 @app.get("/indicatorLayerCombinations")