+++ workers/ohsome_quality_analyst/oqt.py
@@ -13,13 +13,14 @@ import ohsome_quality_analyst.geodatabase.client as db_client
 from ohsome_quality_analyst.base.indicator import BaseIndicator
 from ohsome_quality_analyst.base.report import BaseReport
 from ohsome_quality_analyst.utils.definitions import GEOM_SIZE_LIMIT, INDICATOR_LAYER
+from ohsome_quality_analyst.utils.exceptions import SizeRestrictionError
 from ohsome_quality_analyst.utils.helper import loads_geojson, name_to_class
 
 
 async def create_indicator_as_geojson(
     name: str,
     layer_name: str,
-    bpolys: Optional[str] = None,
+    bpolys: Optional[dict] = None,
     dataset: Optional[str] = None,
     feature_id: Optional[str] = None,
     fid_field: Optional[str] = None,
@@ -34,7 +35,11 @@ async def create_indicator_as_geojson(
     if bpolys is not None:
         features = []
         for i, feature in enumerate(loads_geojson(bpolys)):
-            logging.info("Input feature index: " + str(i))
+            if "id" in feature.keys():
+                id_ = str(feature["id"])
+            else:
+                id_ = str(i)
+            logging.info("Input feature identifier: " + id_)
             if size_restriction:
                 await check_area_size(feature.geometry)
             indicator = await create_indicator(
@@ -67,7 +72,7 @@ async def create_indicator_as_geojson(
 
 async def create_report_as_geojson(
     name: str,
-    bpolys: Optional[str] = None,
+    bpolys: Optional[dict] = None,
     dataset: Optional[str] = None,
     feature_id: Optional[str] = None,
     fid_field: Optional[str] = None,
@@ -82,7 +87,11 @@ async def create_report_as_geojson(
     if bpolys is not None:
         features = []
         for i, feature in enumerate(loads_geojson(bpolys)):
-            logging.info("Input feature index: " + str(i))
+            if "id" in feature.keys():
+                id_ = str(feature["id"])
+            else:
+                id_ = str(i)
+            logging.info("Input feature identifier: " + id_)
             if size_restriction:
                 await check_area_size(feature.geometry)
             report = await create_report(
@@ -265,7 +274,7 @@ async def create_all_indicators(force: bool = False) -> None:
 
 async def check_area_size(geom: Union[Polygon, MultiPolygon]):
     if await db_client.get_area_of_bpolys(geom) > GEOM_SIZE_LIMIT:
-        raise ValueError(
+        raise SizeRestrictionError(
             "Input GeoJSON Object is too big. "
             "The area should be less than {0} km².".format(GEOM_SIZE_LIMIT)
         )