+++ workers/tests/integrationtests/test_api_indicator.py
@@ -1,10 +1,11 @@
 """
 Testing FastAPI Applications:
 https://fastapi.tiangolo.com/tutorial/testing/
+
+For tests regarding the `bpolys` parameter see `test_api_indicator_geojson_io.py`.
 """
-import os
 import unittest
-from typing import Optional
+from urllib.parse import urlencode
 
 from fastapi.testclient import TestClient
 from schema import Schema
@@ -13,6 +14,8 @@ from ohsome_quality_analyst.api.api import app
 
 from .api_response_schema import get_general_schema, get_indicator_feature_schema
 from .utils import oqt_vcr
+
+ENDPOINT = "/indicator"
 
 
 class TestApiIndicator(unittest.TestCase):
@@ -31,6 +34,7 @@ class TestApiIndicator(unittest.TestCase):
 
     def run_tests(self, response) -> None:
         self.assertEqual(response.status_code, 200)
+        self.assertEqual(response.headers["content-type"], "application/geo+json")
         for schema in (self.general_schema, self.feature_schema):
             self.validate(response.json(), schema)
 
@@ -38,134 +42,103 @@ class TestApiIndicator(unittest.TestCase):
         schema.validate(geojson)  # Print information if validation fails
         self.assertTrue(schema.is_valid(geojson))
 
-    def get_response(
-        self,
-        indicator_name: str,
-        layer_name: str,
-        dataset: str,
-        feature_id: str,
-        fid_field: Optional[str] = None,
-    ):
-        """Return HTTP GET response"""
-        if fid_field is not None:
-            base_url = "/indicator/{0}?".format(self.indicator_name)
-            parameter = "layerName={0}&dataset={1}&featureId={2}&fidField={3}".format(
-                layer_name,
-                dataset,
-                feature_id,
-                fid_field,
-            )
-            url = base_url + parameter
-        else:
-            url = "/indicator/{0}?layerName={1}&dataset={2}&featureId={3}".format(
-                indicator_name,
-                layer_name,
-                dataset,
-                feature_id,
-            )
-        return self.client.get(url)
-
-    def post_response(
-        self,
-        indicator_name: str,
-        layer_name: str,
-        dataset: str,
-        feature_id: str,
-        fid_field: Optional[str] = None,
-    ):
-        """Return HTTP POST response"""
-        if fid_field is not None:
-            data = {
-                "dataset": self.dataset,
-                "featureId": self.feature_id,
-                "layerName": self.layer_name,
-            }
-        else:
-            data = {
-                "dataset": dataset,
-                "featureId": feature_id,
-                "layerName": layer_name,
-                "fidField": fid_field,
-            }
-        url = "/indicator/{0}".format(indicator_name)
-        return self.client.post(url, json=data)
-
     @oqt_vcr.use_cassette()
     def test_indicator_dataset_default_fid_field(self):
-        parameters = (
-            self.indicator_name,
-            self.layer_name,
-            self.dataset,
-            self.feature_id,
-        )
+        parameters = {
+            "name": self.indicator_name,
+            "layerName": self.layer_name,
+            "dataset": self.dataset,
+            "featureId": self.feature_id,
+        }
         for response in (
-            self.get_response(*parameters),
-            self.post_response(*parameters),
+            self.client.get(ENDPOINT + "?" + urlencode(parameters)),
+            self.client.post(ENDPOINT, json=parameters),
         ):
             self.run_tests(response)
 
     @oqt_vcr.use_cassette()
     def test_indicator_dataset_custom_fid_field(self):
-        parameters = (
-            self.indicator_name,
-            self.layer_name,
-            self.dataset,
-            self.feature_id,
-            self.fid_field,
-        )
+        parameters = {
+            "name": self.indicator_name,
+            "layerName": self.layer_name,
+            "dataset": self.dataset,
+            "featureId": self.feature_id,
+            "fidField": self.fid_field,
+        }
         for response in (
-            self.get_response(*parameters),
-            self.post_response(*parameters),
+            self.client.get(ENDPOINT + "?" + urlencode(parameters)),
+            self.client.post(ENDPOINT, json=parameters),
         ):
             self.run_tests(response)
 
     @oqt_vcr.use_cassette()
     def test_indicator_dataset_custom_fid_field_2(self):
-        parameters = (
-            self.indicator_name,
-            self.layer_name,
-            self.dataset,
-            "Heidelberg",
-            "name",
-        )
+        parameters = {
+            "name": self.indicator_name,
+            "layerName": self.layer_name,
+            "dataset": self.dataset,
+            "featureId": "Heidelberg",
+            "fidField": "name",
+        }
         for response in (
-            self.get_response(*parameters),
-            self.post_response(*parameters),
+            self.client.get(ENDPOINT + "?" + urlencode(parameters)),
+            self.client.post(ENDPOINT, json=parameters),
         ):
             self.run_tests(response)
 
     @oqt_vcr.use_cassette()
     def test_indicator_dataset_invalid(self):
-        data = {
+        parameters = {
+            "name": self.indicator_name,
+            "layerName": self.layer_name,
             "dataset": "foo",
-            "featureId": "3",
+            "featureId": self.feature_id,
         }
-        url = "/indicator/{0}".format(self.indicator_name)
-        response = self.client.post(url, json=data)
-        self.assertEqual(response.status_code, 422)
+        for response in (
+            self.client.get(ENDPOINT + "?" + urlencode(parameters)),
+            self.client.post(ENDPOINT, json=parameters),
+        ):
+            self.assertEqual(response.status_code, 422)
+            content = response.json()
+            self.assertEqual(content["type"], "RequestValidationError")
 
     @oqt_vcr.use_cassette()
     def test_indicator_invalid_set_of_arguments(self):
-
-        path = os.path.join(
-            os.path.dirname(os.path.abspath(__file__)),
-            "fixtures",
-            "heidelberg-altstadt-feature.geojson",
-        )
-        with open(path, "r") as f:
-            bpolys = f.read()
-        url = "/indicator/{0}".format(self.indicator_name)
-        for data in (
+        for parameters in (
             {
-                "bpolys": bpolys,
-                "dataset": "foo",
-                "featureId": "3",
+                "name": self.indicator_name,
+                "layerName": "building_count",
+                "dataset": "regions",
             },
-            {"dataset": "regions"},
-            {"feature_id": "3"},
+            {
+                "name": self.indicator_name,
+                "layerName": "building_count",
+                "feature_id": "3",
+            },
         ):
-            response = self.client.post(url, json=data)
+            for response in (
+                self.client.get(ENDPOINT + "?" + urlencode(parameters)),
+                self.client.post(ENDPOINT, json=parameters),
+            ):
+                self.assertEqual(response.status_code, 422)
+                content = response.json()
+                self.assertEqual(content["type"], "RequestValidationError")
+
+    @oqt_vcr.use_cassette()
+    def test_indicator_invalid_layer(self):
+        parameters = {
+            "name": self.indicator_name,
+            "layerName": "amenities",
+            "dataset": "regions",
+            "featureId": "3",
+        }
+        for response in (
+            self.client.get(ENDPOINT + "?" + urlencode(parameters)),
+            self.client.post(ENDPOINT, json=parameters),
+        ):
             self.assertEqual(response.status_code, 422)
+            content = response.json()
+            self.assertEqual(content["type"], "RequestValidationError")
 
 
 if __name__ == "__main__":