+++ workers/tests/integrationtests/test_api_indicator_geojson_io.py
@@ -2,10 +2,10 @@
 Testing FastAPI Applications:
 https://fastapi.tiangolo.com/tutorial/testing/
 """
-
 import os
 import unittest
 from typing import Tuple
+from urllib.parse import urlencode
 
 from fastapi.testclient import TestClient
 from schema import Schema
@@ -17,13 +17,7 @@ from .api_response_schema import (
     get_general_schema,
     get_indicator_feature_schema,
 )
-from .utils import oqt_vcr
-
-
-def get_fixture(name):
-    path = os.path.join(os.path.dirname(os.path.abspath(__file__)), "fixtures", name)
-    with open(path, "r") as f:
-        return f.read()
+from .utils import get_geojson_fixture, oqt_vcr
 
 
 class TestApiIndicatorIo(unittest.TestCase):
@@ -39,43 +33,49 @@ class TestApiIndicatorIo(unittest.TestCase):
 
     def run_tests(self, response, schemata: Tuple[Schema]) -> None:
         self.assertEqual(response.status_code, 200)
+        self.assertEqual(response.headers["content-type"], "application/geo+json")
         for schema in schemata:
             self.validate(response.json(), schema)
 
-    def validate(self, geojson: dict, schema: Schema) -> None:
-        schema.validate(geojson)  # Print information if validation fails
-        self.assertTrue(schema.is_valid(geojson))
+    def validate(self, geojson_: dict, schema: Schema) -> None:
+        schema.validate(geojson_)  # Print information if validation fails
+        self.assertTrue(schema.is_valid(geojson_))
 
     def get_response(self, bpoly):
         """Return HTTP GET response"""
-        url = "/indicator/{0}?layerName={1}&bpolys={2}".format(
-            self.indicator_name,
-            self.layer_name,
-            bpoly,
+        parameters = urlencode(
+            {
+                "name": self.indicator_name,
+                "layerName": self.layer_name,
+                "bpolys": bpoly,
+            }
         )
-        return self.client.get(url)
+        return self.client.get("/indicator?" + parameters)
 
     def post_response(self, bpoly):
         """Return HTTP POST response"""
-        data = {"bpolys": bpoly, "layerName": self.layer_name}
-        url = f"/indicator/{self.indicator_name}"
-        return self.client.post(url, json=data)
+        parameters = {
+            "name": self.indicator_name,
+            "bpolys": bpoly,
+            "layerName": self.layer_name,
+        }
+        return self.client.post("/indicator", json=parameters)
 
     @oqt_vcr.use_cassette()
     def test_indicator_bpolys_geometry(self):
-        geometry = get_fixture("heidelberg-altstadt-geometry.geojson")
+        geometry = get_geojson_fixture("heidelberg-altstadt-geometry.geojson")
         for response in (self.get_response(geometry), self.post_response(geometry)):
             self.run_tests(response, (self.general_schema, self.feature_schema))
 
     @oqt_vcr.use_cassette()
     def test_indicator_bpolys_feature(self):
-        feature = get_fixture("heidelberg-altstadt-feature.geojson")
+        feature = get_geojson_fixture("heidelberg-altstadt-feature.geojson")
         for response in (self.get_response(feature), self.post_response(feature)):
             self.run_tests(response, (self.general_schema, self.feature_schema))
 
     @oqt_vcr.use_cassette()
     def test_indicator_bpolys_featurecollection(self):
-        featurecollection = get_fixture(
+        featurecollection = get_geojson_fixture(
             "heidelberg-bahnstadt-bergheim-featurecollection.geojson",
         )
         for response in (
@@ -91,25 +91,38 @@ class TestApiIndicatorIo(unittest.TestCase):
 
     @oqt_vcr.use_cassette()
     def test_indicator_bpolys_size_limit(self):
-        feature = get_fixture("europe.geojson")
-        with self.assertRaises(ValueError):
-            self.get_response(feature)
-        with self.assertRaises(ValueError):
-            self.post_response(feature)
+        feature = get_geojson_fixture("europe.geojson")
+        for response in (
+            self.get_response(feature),
+            self.post_response(feature),
+        ):
+            self.assertEqual(response.status_code, 422)
+            content = response.json()
+            self.assertEqual(content["type"], "SizeRestrictionError")
 
     @oqt_vcr.use_cassette()
-    def test_bpolys_invalid(self):
+    def test_invalid_set_of_arguments(self):
         path = os.path.join(
             os.path.dirname(os.path.abspath(__file__)),
             "fixtures",
-            "invalid.geojson",
+            "heidelberg-altstadt-feature.geojson",
         )
-        with open(path, "r") as file:
-            bpolys = file.read()
-        data = {"bpolys": bpolys, "layerName": self.layer_name}
-        url = f"/indicator/{self.indicator_name}"
-        response = self.client.post(url, json=data)
-        self.assertEqual(response.status_code, 422)
+        with open(path, "r") as f:
+            bpolys = f.read()
+        parameters = {
+            "name": self.indicator_name,
+            "bpolys": bpolys,
+            "dataset": "foo",
+            "featureId": "3",
+        }
+
+        for response in (
+            self.client.get("/indicator?" + urlencode(parameters)),
+            self.client.post("/indicator", json=parameters),
+        ):
+            self.assertEqual(response.status_code, 422)
+            content = response.json()
+            self.assertEqual(content["type"], "RequestValidationError")
 
 
 if __name__ == "__main__":