+++ workers/tests/integrationtests/test_api_report_geojson_io.py
@@ -2,9 +2,8 @@
 Testing FastAPI Applications:
 https://fastapi.tiangolo.com/tutorial/testing/
 """
-import os
 import unittest
-import urllib.parse
+from urllib.parse import urlencode
 
 import geojson
 from fastapi.testclient import TestClient
@@ -17,13 +16,7 @@ from .api_response_schema import (
     get_general_schema,
     get_report_feature_schema,
 )
-from .utils import oqt_vcr
-
-
-def get_fixture(name):
-    path = os.path.join(os.path.dirname(os.path.abspath(__file__)), "fixtures", name)
-    with open(path, "r") as f:
-        return f.read()
+from .utils import get_geojson_fixture, oqt_vcr
 
 
 class TestApiReportIo(unittest.TestCase):
@@ -46,6 +39,7 @@ class TestApiReportIo(unittest.TestCase):
 
     def run_tests(self, response) -> None:
         self.assertEqual(response.status_code, 200)
+        self.assertEqual(response.headers["content-type"], "application/geo+json")
 
         response_content = geojson.loads(response.content)
         self.assertTrue(response_content.is_valid)  # Valid GeoJSON?
@@ -54,33 +48,31 @@ class TestApiReportIo(unittest.TestCase):
 
     def get_response(self, bpoly):
         """Return HTTP GET response"""
-        parameters = urllib.parse.urlencode({"bpolys": bpoly})
-        url = "/report/{0}?{1}".format(
-            self.report_name,
-            parameters,
-        )
+        parameters = urlencode({"name": self.report_name, "bpolys": bpoly})
+        url = "/report?" + parameters
         return self.client.get(url)
 
     def post_response(self, bpoly):
         """Return HTTP POST response"""
-        url = "/report/{0}".format(self.report_name)
-        return self.client.post(url, json={"bpolys": bpoly})
+        data = {"name": self.report_name, "bpolys": bpoly}
+        url = "/report"
+        return self.client.post(url, json=data)
 
     @oqt_vcr.use_cassette()
     def test_report_bpolys_geometry(self):
-        geometry = get_fixture("heidelberg-altstadt-geometry.geojson")
+        geometry = get_geojson_fixture("heidelberg-altstadt-geometry.geojson")
         for response in (self.get_response(geometry), self.post_response(geometry)):
             self.run_tests(response)
 
     @oqt_vcr.use_cassette()
     def test_report_bpolys_feature(self):
-        feature = get_fixture("heidelberg-altstadt-feature.geojson")
+        feature = get_geojson_fixture("heidelberg-altstadt-feature.geojson")
         for response in (self.get_response(feature), self.post_response(feature)):
             self.run_tests(response)
 
     @oqt_vcr.use_cassette()
     def test_report_bpolys_featurecollection(self):
-        featurecollection = get_fixture(
+        featurecollection = get_geojson_fixture(
             "heidelberg-bahnstadt-bergheim-featurecollection.geojson",
         )
         for response in (
@@ -101,11 +93,29 @@ class TestApiReportIo(unittest.TestCase):
 
     @oqt_vcr.use_cassette()
     def test_report_bpolys_size_limit(self):
-        feature = get_fixture("europe.geojson")
-        with self.assertRaises(ValueError):
-            self.get_response(feature)
-        with self.assertRaises(ValueError):
-            self.post_response(feature)
+        feature = get_geojson_fixture("europe.geojson")
+        response = self.post_response(feature)
+        self.assertEqual(response.status_code, 422)
+
+    @oqt_vcr.use_cassette()
+    def test_invalid_set_of_arguments(self):
+        bpolys = get_geojson_fixture("heidelberg-altstadt-feature.geojson")
+        parameters = {
+            "name": self.report_name,
+            "bpolys": bpolys,
+            "dataset": "foo",
+            "featureId": "3",
+        }
+
+        response = self.client.get("/report?" + urlencode(parameters))
+        self.assertEqual(response.status_code, 422)
+        content = response.json()
+        self.assertEqual(content["type"], "RequestValidationError")
+
+        response = self.client.post("/report", json=parameters)
+        self.assertEqual(response.status_code, 422)
+        content = response.json()
+        self.assertEqual(content["type"], "RequestValidationError")
 
 
 if __name__ == "__main__":